libprobe: Add a probe! macro for static instrumentation
This adds a new libprobe with a probe! macro, for adding static
instrumentation to any rust crate.  Please see the docstrings for simple
probe examples and how they might be used by tools.

This initial version only implements SystemTap SDT, with a voided macro
for other platforms, so probes can be freely written anywhere.  I
believe the macro is generic enough that a DTrace USDT probe! could be
implemented for freebsd and macos too.