systemtap: add SHF_GNU_RETAIN flag to .stapsdt.base
Gold, lld, and mold strip the .stapsdt.base section when invoked with the
--gc-sections option, which is appears to be commonly enabled in regular
builds.
GNU bfd linker seems to be the only one among the tested linker to preserve
the section as it seems to apply less aggressive criteria for sections GC.

This patch adds the SHF_GNU_RETAIN flag to the .stapsdt.base section, making
it a GC-root. This ensures the section is always retained across different
linker implementations.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>